+++ extensions/music/src/reveal-track.tsx
@@ -4,8 +4,9 @@ import * as TE from "fp-ts/TaskEither";
 import * as music from "./util/scripts";
 import { handleTaskEitherError } from "./util/utils";
 
+// Run both commands ignoring activate error
 export default pipe(
-  music.currentTrack.love,
-  TE.chain(() => pipe(music.currentTrack.addToLibrary)),
-  handleTaskEitherError("Failed to save/love the track", "✅ Track loved & added to library")
+  music.currentTrack.reveal,
+  TE.apFirst(music.general.activate),
+  handleTaskEitherError("Could not display current track")
 )();