+++ extensions/music/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Apple Music Changelog
 
+## [Update] - 2022-10-29
+
+- Fixed "Add to Library" script.
+- Improved error handling
+- Added "Reveal Track" command
+
 ## [Update] - 2022-07-27
+
 - Improved error reporting
 - New README and screenshots
 
@@ -16,6 +23,7 @@ Squashed a few bugs in "Play Library Album"
 - Renamed "Search & Play Library Track" to "Play Library Track"
 
 ## [Fix] - 2022-05-30
+
 Fixed issue that caused the extension to crash
 
 ## [Track Rating / Fixes] - 2022-05-05